backend: move http handler to transform pkg
