+++ backend/pkg/api/connect/integration/transform_test.go
@@ -16,7 +16,7 @@ import (
 	rqr "github.com/stretchr/testify/require"
 	"github.com/twmb/franz-go/pkg/kadm"
 
-	"github.com/redpanda-data/console/backend/pkg/api"
+	"github.com/redpanda-data/console/backend/pkg/api/connect/service/transform"
 	v1alpha1 "github.com/redpanda-data/console/backend/pkg/protogen/redpanda/api/dataplane/v1alpha1"
 	v1alpha1connect "github.com/redpanda-data/console/backend/pkg/protogen/redpanda/api/dataplane/v1alpha1/dataplanev1alpha1connect"
 )
@@ -91,7 +91,7 @@ func (s *APISuite) TestDeployTransform() {
 	body := &bytes.Buffer{}
 	writer := multipart.NewWriter(body)
 
-	metadataOut := api.TransformMetadata{
+	metadataOut := transform.Metadata{
 		Name:         tfName,
 		InputTopic:   inputTopicName,
 		OutputTopics: []string{outputTopicName},
